test(runner): verify promotion notification after destroy panic
