+++ crates/runner/src/cmd/start/tests/idle_reuse/workspace_cache.rs
@@ -3,7 +3,7 @@ use super::super::support::{
     WorkspacePromotionSeedSpec, context_with_session, mock_run_config,
     mock_run_config_with_overrides, push_job, seed_idle_pool_with_overrides,
     seed_idle_pool_with_workspace_promotion, shutdown, test_profiles, wait_budget_count,
-    wait_discover_entered, wait_heartbeat_with_workspace_after, wait_idle_pool_reuse_keys,
+    wait_discover_entered, wait_idle_pool_reuse_keys,
 };
 
 use crate::idle_reuse_preparation::add_healthy_reuse_preparation_matcher;
@@ -18,6 +18,43 @@ fn reusable_candidate(
 ) -> crate::provider::JobCandidate {
     crate::provider::JobCandidate::new(run_id, profile_name.to_string())
         .with_reuse_key(Some(reuse_key.to_string()))
+}
+
+async fn wait_heartbeat_with_workspace_after(
+    handle: &crate::provider::mock::MockProviderHandle,
+    mut cursor: usize,
+    reuse_key: &str,
+    expected_workspace: &WorkspaceCacheCapability,
+    absent_sandbox_reuse_key: Option<&str>,
+    timeout: Duration,
+) -> bool {
+    let deadline = tokio::time::Instant::now() + timeout;
+    loop {
+        {
+            let heartbeats = handle.heartbeats.lock().unwrap_or_else(|e| e.into_inner());
+            if heartbeats[cursor..].iter().any(|heartbeat| {
+                let has_expected_workspace = heartbeat.held_workspace_states.iter().any(|state| {
+                    state.reuse_key == reuse_key
+                        && state.workspace_caches.contains(expected_workspace)
+                });
+                let omits_claimed_sandbox = absent_sandbox_reuse_key.is_none_or(|reuse_key| {
+                    heartbeat
+                        .held_sandbox_states
+                        .iter()
+                        .all(|state| state.reuse_key != reuse_key)
+                });
+                has_expected_workspace && omits_claimed_sandbox
+            }) {
+                return true;
+            }
+            cursor = heartbeats.len();
+        }
+
+        let remaining = deadline.saturating_duration_since(tokio::time::Instant::now());
+        if remaining.is_zero() || !handle.wait_heartbeat_past(cursor, remaining).await {
+            return false;
+        }
+    }
 }
 
 #[tokio::test]
