+++ crates/runner/src/cmd/start/tests/support/mod.rs
@@ -26,7 +26,7 @@ pub(super) use self::status::{
 pub(super) use self::wait::{
     assert_run_exits_within, wait_budget_count, wait_budget_exhausted_reactor, wait_cancel_handle,
     wait_cancel_token, wait_cancel_token_removed, wait_discover_entered,
-    wait_heartbeat_with_workspace_after, wait_idle_cleanup_processed_with_expired_entries,
-    wait_idle_pool_len, wait_idle_pool_reuse_keys, wait_parking_state,
-    wait_sandbox_lifecycle_counts, wait_usage_flush_requested, wait_workspace_cache_reuse_keys,
+    wait_idle_cleanup_processed_with_expired_entries, wait_idle_pool_len,
+    wait_idle_pool_reuse_keys, wait_parking_state, wait_sandbox_lifecycle_counts,
+    wait_usage_flush_requested, wait_workspace_cache_reuse_keys,
 };
